tests: add logging of nodes logs in 'tmp' directory
Quite often is needed to have the realtime logs of several nodes in
order to investigate the problem, redirect the node output to the
terminal might help only with one node, in this commit multiwriter was
used in order to copy the standart and error outputs in the several
places. With debug field specified to 'true' now node will create the
log file inside the '/tmp/' directory, convention of log file is
following: [name_of_the_node].txt. Name is specified in node
constructor.