channeldb: add index store
IndexStore is the bold db storage where we assign the index to every
object we store in increasing manner, by doing this we preserve the
order of the objects and also give the ability to the user to quickly
retrieve the object by using this index. By using index store subsystem
might initialize the map of the object_id<->index during initialization
and than use it for checking the existence of object without touching
the db itself.

Also in this commit Storable interface have been create which allow the
reusage of the IndexStore with different type of objects.