htlcswicth: start use bounded links in htlc switch
In this commit an attempt have been made to replicate the original idea
of bounded channel which has been applied to earlier version of htlc
switch.

NOTE: Such approach doesn't limits the number of the htlc received from
remote peer, so it means that remote peer still might cause the channel
to be flooded with outstanding htlcs.