[review] htlcswitch: remove bound channel link usage
In order to fix the bug when the balance is changing before we handle
the packet, thereby leading to the negative balance, we need to avoid
calling the channel link packet handler inside the goroutine. We might
did this by restoring the previous approach when we stored channel
link balances inside the switch, but such approach lead to the
unnecessary sychronization and makes channel link to be more coupled
with switch. By removing the bound channel link and introducing the
channel link packet queue we solve this problem.