[review] fix htlc overflow by introducing the reprocessing queue
In this commit in channel link downstream reprocessing queue have been
added, which consume the htlc which have been returned by add htlc
function with max htlc number error. This queue allows collect the
pending htlc in memore (possibly on disk in future) and reprocess it
when new slots will be available (after receiving the settle or fail
messages).