lnwire: add non-spec onion code
As a temporary measure the additional non-spec error have been added in
order to simplify the transition to the usage of the channel ids in the
switch. For now in order to generate the proper error inside the switch
we would need creating additional functions inside the switch config,
which is not needed because after switching to use channel ids the
generation of the error would be moved inside the channel link.